feat: space activity tab reuses explore feed UI
Replaces the legacy minimal activity list with the explore feed cards
— types pills, cover thumbnails, vote buttons, comment count, infinite
scroll. Space is locked to the current space (no space dropdown, no
space link on cards, no Join button); type whitelist and name filter
are both disabled so Properties, Types, unnamed entities, etc. also
surface.

Implementation extracts a generic <EntityFeed> out of <ExplorePage>.
Both surfaces are now thin wrappers; when the Top sort lands on
explore later it flows through to activity for free.

- core/explore/fetch-explore-feed.ts: optional `typeIds` and
  `requireName` args. Preserves existing explore defaults via the
  explore route passing EXPLORE_ENTITY_TYPE_IDS and requireName=true.
- partials/feed/entity-feed.tsx: new generic feed (time dropdown,
  optional space dropdown, auth-aware infinite scroll, skeletons).
- partials/explore/explore-page.tsx: thin wrapper around EntityFeed.
- partials/explore/explore-feed-card.tsx: `hideSpaceLink` and
  `hideJoinButton` props (default false, preserves explore behavior).
- app/api/activity/feed/route.ts: new activity-specific endpoint that
  calls fetchExploreFeed with typeIds=undefined, requireName=false.
- partials/activity/activity-feed-page.tsx: new thin client wrapper.
- app/space/[id]/(space)/activity/page.tsx: renders ActivityFeedPage.
- Deleted: partials/activity/activity.tsx,
  partials/activity/activity-server-container.tsx.
