+++ apps/web/app/space/[id]/(space)/activity/page.tsx
@@ -1,12 +1,8 @@
 import { IdUtils } from '@geoprotocol/geo-sdk/lite';
 
-import { Suspense } from 'react';
-
 import { notFound } from 'next/navigation';
 
-import { Skeleton } from '~/design-system/skeleton';
-
-import { ActivityServerContainer } from '~/partials/activity/activity-server-container';
+import { ActivityFeedPage } from '~/partials/activity/activity-feed-page';
 
 interface Props {
   params: Promise<{ id: string }>;
@@ -19,28 +15,5 @@ export default async function Activity(props: Props) {
     notFound();
   }
 
-  return (
-    <Suspense fallback={<ActivitySkeleton />}>
-      <ActivityServerContainer spaceId={params.id} />
-    </Suspense>
-  );
+  return <ActivityFeedPage spaceId={params.id} />;
 }
-
-const ActivitySkeleton = () => {
-  return (
-    <div className="divide-y divide-divider">
-      {new Array(3).fill(0).map((_, index) => (
-        <div key={index} className="flex items-center gap-5 py-4">
-          <div>
-            <Skeleton className="size-10 rounded-md" />
-          </div>
-          <div className="flex-1">
-            <Skeleton className="h-[21px] w-1/2" />
-            <Skeleton className="mt-1 h-[32px] w-3/4" />
-            <Skeleton className="mt-3 h-[15px] w-1/4" />
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-};
