+++ apps/web/partials/activity/activity-feed-page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { EntityFeed } from '~/partials/feed/entity-feed';
+
+export function ActivityFeedPage({ spaceId }: { spaceId: string }) {
+  return <EntityFeed apiEndpoint="/api/activity/feed" lockedSpaceId={spaceId} />;
+}
